pyproject.toml
[build-system]
requires = ["setuptools>=61.0", 'wheel']
build-backend = "setuptools.build_meta"

[project]
name = "uiucprescon.build"
description = "Builder for python projects that require C or C++ dependencies or bindings"
classifiers = [
    "License :: OSI Approved :: University of Illinois/NCSA Open Source License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = ["version"]

dependencies = [
    "setuptools>=61.0",
    'wheel',
    "conan>=1.53,!=1.55.0",
    "cmake",
    'pybind11~=2.10.1',
    "conan>=1.53,!=1.55.0,<2.0",
    'toml'
]
license = {text = "University of Illinois/NCSA Open Source License"}
readme = "README.md"
requires-python = ">=3.7"
[tool.setuptools.dynamic]
version = {attr = "uiucprescon.build.VERSION"}

[tool.setuptools]
packages = ['uiucprescon.build']

[tool.coverage.run]
relative_files = false
branch = true
source = [
    ".",
#    "uiucprescon/build"
]
#[tool.coverage.paths]
[tool.coverage.report]
source = [
    ".",
    "uiucprescon/build"
]
include_namespace_packages= true
[tool.coverage.paths]

#[tool.coverage.report]
